extract enemies && player
