+++ frontend/src/Components/phaser/enemies/LadyBugSpawner.js
